handled community creation and fetching and unifying DTOs, handling issues with false redirection when trying to sign in with google, altered the services to match the backend and checked the CORS header on the backend routes

+++ Backend/src/main/java/com/planchella/utils/IdGenerator.java
@@ -7,6 +7,6 @@ public class IdGenerator {
 
     public static Long generateId() {
         // Generate a positive long ID (excluding 0)
-        return Math.abs(random.nextLong()) % (Long.MAX_VALUE - 1) + 1;
+        return Math.abs(random.nextLong()) % (Integer.MAX_VALUE - 1) + 1;
     }
 }
